requirements.txt
cryptography>=42.0.0
liboqs-python==0.15.0
